Support using the real event store with the command testing DSL
Unifies the real and fake event store APIs so the same testing
DSL (`given_events`, `when_command`, `then_events`) can be used with
both event stores.

Test correct support for aggregate unique keys using this testing DSL.
